third_party/WebKit/LayoutTests/http/tests/preload/dynamic_adding_preload.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that dynamically added preloaded resource is downloaded');
</script>
<body>
<script>
    var link = document.createElement("link");
    link.as = "script";
    link.rel = "preload";
    link.href = "../resources/dummy.js";
    document.body.appendChild(link);
    window.addEventListener("load", t.step_func(function() {
        assert_equals(performance.getEntriesByType("resource").length, 4);
        t.done();
    }));
</script>
<script src="../resources/slow-script.pl?delay=200"></script>
</body>
